generators/gemini_generator.py
import os
import logging
import google.generativeai as genai
from PIL import Image
from dotenv import load_dotenv
import io
import requests # Need requests to fetch image data if URL is returned

from .base import BaseImageGenerator

logger = logging.getLogger(__name__)

# Placeholder for potential future Gemini image generation models via google-generativeai
# Currently (as of early 2024), direct text-to-image generation might not be available
# via this specific library/API endpoint. Imagen on Vertex AI is the typical route.
# Update this model name if/when a suitable one becomes available.
GEMINI_IMAGE_MODEL = "gemini-1.5-flash" # Or a future image model

class GeminiImageGenerator(BaseImageGenerator):
    """Image generator using Google's Gemini models (via google-generativeai)."""

    def __init__(self, api_key: str = None, model: str = GEMINI_IMAGE_MODEL):
        load_dotenv() # Load from .env if api_key is not provided directly
        self.api_key = api_key or os.getenv('GEMINI_API_KEY')
        if not self.api_key:
            raise ValueError("Gemini API key not found. Set GEMINI_API_KEY environment variable or pass it directly.")
        
        try:
            genai.configure(api_key=self.api_key)
            self.model = genai.GenerativeModel(model)
            logger.info(f"Gemini Image Generator configured with model: {model}")
            # Check if the model actually supports image generation (this check might need refinement)
            # For now, we proceed and let generate_image handle limitations.
        except Exception as e:
             logger.error(f"Failed to configure Gemini client: {e}")
             raise ValueError(f"Failed to configure Gemini client: {e}") from e

    def generate_image(self, prompt: str) -> Image.Image:
        """Generates an image using the Gemini API (if supported)."""
        logger.info(f"🤖 Generating image with Gemini ({self.model.model_name})...")
        logger.debug(f"Prompt: {prompt[:100]}...")

        # --- IMPORTANT LIMITATION --- 
        # As of the current understanding of the google-generativeai library,
        # direct text-to-image generation like DALL-E is typically handled by 
        # Imagen on Vertex AI, not the standard Gemini API models (like gemini-pro).
        # The code below attempts generation but might fail or require a specific
        # image generation model name if/when available through this API.
        # We will raise NotImplementedError for now.
        # -----------------------------
        
        # Placeholder for actual generation logic if API evolves
        # try:
        #     # Example hypothetical call (syntax might differ)
        #     response = self.model.generate_content(
        #         prompt, 
        #         generation_config=genai.types.GenerationConfig(
        #             # candiate_count=1, # Example config
        #             # temperature=0.9
        #         )
        #         # Potentially specify response mime type if needed
        #     )

        #     # Check response type and extract image data
        #     # This part is highly speculative as the API might return URLs, bytes, etc.
        #     if hasattr(response, 'parts') and response.parts:
        #          image_part = next((part for part in response.parts if 'image' in part.mime_type), None)
        #          if image_part and hasattr(image_part, 'data'): # Assuming direct bytes
        #              image_data = image_part.data
        #              image = Image.open(io.BytesIO(image_data))
        #              logger.info(f"✅ Image generated successfully by Gemini ({image.size})")
        #              return image
        #          elif image_part and hasattr(image_part, 'file_data') and image_part.file_data.file_uri: # Assuming URL
        #             image_url = image_part.file_data.file_uri
        #             logger.info(f"Fetching image from Gemini URL: {image_url}")
        #             img_response = requests.get(image_url)
        #             img_response.raise_for_status()
        #             image = Image.open(io.BytesIO(img_response.content))
        #             logger.info(f"✅ Image generated successfully by Gemini ({image.size})")
        #             return image
                 
        #     raise Exception("Respuesta inválida de Gemini: No se encontró imagen generada.")

        # except Exception as e:
        #     logger.error(f"❌ Error inesperado durante la generación con Gemini: {str(e)}")
        #     import traceback
        #     logger.error(f"Traceback: {traceback.format_exc()}")
        #     raise Exception(f"Error generando imagen con Gemini: {str(e)}") from e

        logger.error("Direct text-to-image generation is not currently supported via the standard Gemini API and google-generativeai library. Use Imagen on Vertex AI.")
        raise NotImplementedError("Image generation via google-generativeai for Gemini is not implemented. Use Imagen on Vertex AI.")
