requirements.txt
openai==1.75.0
python-dotenv==1.1.0
Pillow==11.2.1
numpy==2.2.4
PyYAML==6.0.2
scipy==1.15.2
google-generativeai==0.8.5
google-cloud-aiplatform==1.89.0
requests>=2.32.3
qdrant-client==1.13.3
termcolor==3.0.1